+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 
 ## 0.8.0.0 -- Unreleased
 
-* New builtin: `\lastValue`, `\unmul`.
+* New builtin: `\lastValue`, `\shortestValue`, `\unmul`, `\unsnoc0`.
 * A new mode `last` is added. It outputs the last possible result. You can switch to this mode by the `-t` flag in the command line or the `\Mode last` command in the REPL.
 * Now input can have leading and trailing spaces. The interpreter will trim them when parsing.
 * Now Nekomata can be built with GHC 9.10.1, thanks to an update of the `haskeline` package.
+* Fixed a bug in the REPL. Now the REPL can correctly handle strings with whitespaces.
 
 ### Breaking changes